refactoring: WIP
