+++ bindings/src/erc20_forwarder.rs
@@ -1,0 +1,310 @@
+use crate::permit2::permit_witness_transfer_from_signature;
+use alloy::hex;
+use alloy::primitives::{address, Address, B256, U256};
+use alloy::signers::local::PrivateKeySigner;
+use arm_risc0::action_tree::MerkleTree;
+use arm_risc0::authorization::{AuthorizationSigningKey, AuthorizationVerifyingKey};
+use arm_risc0::compliance::INITIAL_ROOT;
+use arm_risc0::encryption::{random_keypair, AffinePoint};
+use arm_risc0::evm::CallType;
+use arm_risc0::merkle_path::MerklePath;
+use arm_risc0::nullifier_key::{NullifierKey, NullifierKeyCommitment};
+
+use crate::conversion::ProtocolAdapter;
+use arm_risc0::utils::bytes_to_words;
+use sha2::{Digest, Sha256};
+use simple_transfer_app::mint::construct_mint_tx;
+use simple_transfer_app::resource::{construct_ephemeral_resource, construct_persistent_resource};
+use simple_transfer_app::transfer::construct_transfer_tx;
+use simple_transfer_app::utils::authorize_the_action;
+
+pub struct SetUp {
+    pub signer: PrivateKeySigner,
+    pub erc20: Address,
+    pub amount: U256,
+    pub nonce: U256,
+    pub deadline: U256,
+    pub spender: Address,
+}
+
+pub fn default_values() -> SetUp {
+    SetUp {
+        signer: "0x97ecae11e1bd9b504ff977ae3815599331c6b0757ee4af3140fe616adb19ae45"
+            .parse()
+            .expect("should parse private key"),
+        erc20: address!("0x5615dEB798BB3E4dFa0139dFa1b3D433Cc23b72f"),
+        amount: U256::from(1000),
+        nonce: U256::from(0),
+        deadline: U256::from(1789040701),
+        spender: address!("0xA4AD4f68d0b91CFD19687c881e50f3A00242828c"),
+    }
+}
+
+pub struct KeyChain {
+    auth_signing_key: AuthorizationSigningKey,
+    nf_key: NullifierKey,
+    discovery_pk: AffinePoint,
+    encryption_pk: AffinePoint,
+}
+
+impl KeyChain {
+    fn auth_verifying_key(&self) -> AuthorizationVerifyingKey {
+        AuthorizationVerifyingKey::from_signing_key(&self.auth_signing_key)
+    }
+
+    fn nullifier(&self, resource: arm_risc0::resource::Resource) -> arm_risc0::Digest {
+        resource.nullifier(&self.nf_key).unwrap()
+    }
+
+    fn nullifier_key_commitment(&self) -> NullifierKeyCommitment {
+        self.nf_key.commit()
+    }
+}
+
+fn keychain_a() -> KeyChain {
+    KeyChain {
+        auth_signing_key: AuthorizationSigningKey::from_bytes(&vec![15u8; 32]),
+        nf_key: NullifierKey::from_bytes(&vec![13u8; 32]),
+        discovery_pk: AffinePoint::GENERATOR,
+        encryption_pk: AffinePoint::GENERATOR,
+    }
+}
+
+fn mint_tx(d: &SetUp, keychain: &KeyChain) -> ProtocolAdapter::Transaction {
+    let latest_cm_tree_root = INITIAL_ROOT.as_words().to_vec();
+
+    let (consumed_resource, created_resource) = mint_consumed_created_pair(d, keychain);
+
+    let action_tree_root = sha256(
+        consumed_resource
+            .nullifier(&keychain.nf_key)
+            .unwrap()
+            .as_bytes(),
+        created_resource.commitment().as_bytes(),
+    );
+
+    let rt = tokio::runtime::Runtime::new().unwrap();
+    let permit_sig = rt.block_on(permit_witness_transfer_from_signature(
+        &d.signer,
+        d.erc20,
+        d.amount,
+        d.nonce,
+        d.deadline,
+        d.spender,
+        action_tree_root, // Witness
+    ));
+
+    // Construct the mint transaction
+    let tx = construct_mint_tx(
+        consumed_resource,
+        latest_cm_tree_root,
+        keychain.nf_key.clone(),
+        keychain.discovery_pk,
+        d.spender.to_vec(),
+        d.erc20.to_vec(),
+        d.signer.address().to_vec(),
+        d.nonce.to_be_bytes_vec(),
+        d.deadline.to_be_bytes_vec(),
+        permit_sig.as_bytes().to_vec(),
+        created_resource,
+        keychain.discovery_pk,
+        keychain.encryption_pk,
+    );
+
+    // Verify the transaction
+    assert!(tx.clone().verify(), "Transaction verification failed");
+
+    ProtocolAdapter::Transaction::from(tx)
+}
+
+fn mint_consumed_created_pair(
+    d: &SetUp,
+    keychain: &KeyChain,
+) -> (arm_risc0::resource::Resource, arm_risc0::resource::Resource) {
+    let consumed_resource = construct_ephemeral_resource(
+        &d.spender.to_vec(),
+        &d.erc20.to_vec(),
+        d.amount.try_into().unwrap(),
+        vec![4u8; 32], // nonce
+        keychain.nf_key.commit(),
+        vec![5u8; 32], // rand_seed
+        CallType::Wrap,
+        &d.signer.address().to_vec(),
+    );
+
+    // ZCash Nonce trick
+    let consumed_nf = consumed_resource.nullifier(&keychain.nf_key).unwrap();
+
+    // Note: We could use a different nullifier key here
+    let created_resource = construct_persistent_resource(
+        &d.spender.to_vec(),
+        &d.erc20.to_vec(),
+        d.amount.try_into().unwrap(),
+        consumed_nf.as_bytes().to_vec(), // nonce
+        keychain.nf_key.commit(),
+        vec![6u8; 32], // rand_seed
+        &keychain.auth_verifying_key(),
+    );
+    (consumed_resource, created_resource)
+}
+
+fn transfer_tx(d: &SetUp, keychain: &KeyChain) -> ProtocolAdapter::Transaction {
+    let (_, minted_resource) = mint_consumed_created_pair(d, &keychain);
+
+    let consumed_nf = minted_resource.nullifier(&keychain.nf_key).unwrap();
+
+    let created_resource = construct_persistent_resource(
+        &d.spender.to_vec(), // forwarder_addr
+        &d.erc20.to_vec(),   // token_addr
+        d.amount.try_into().unwrap(),
+        consumed_nf.as_bytes().to_vec(), // nonce
+        keychain.nf_key.commit(),
+        vec![7u8; 32], // rand_seed
+        &keychain.auth_verifying_key(),
+    );
+    let created_cm = created_resource.commitment();
+
+    // Get the authorization signature, it can be from external signing(e.g. wallet)
+    let action_tree = MerkleTree::new(vec![consumed_nf, created_cm]);
+
+    let auth_sig = authorize_the_action(&keychain.auth_signing_key, &action_tree);
+
+    let empty_hash = B256::from(hex!(
+        "cc1d2f838445db7aec431df9ee8a871f40e7aa5e064fc056633ef8c60fab7b06"
+    ));
+
+    // Construct the transfer transaction
+    let is_left = false;
+    let path: &[(Vec<u32>, bool)] = &[(bytes_to_words(empty_hash.as_slice()), is_left)];
+    let merkle_path = MerklePath::from_path(path);
+
+    let tx = construct_transfer_tx(
+        minted_resource.clone(),
+        merkle_path.clone(),
+        keychain.nf_key.clone(),
+        keychain.auth_verifying_key(),
+        auth_sig,
+        keychain.discovery_pk,
+        keychain.encryption_pk,
+        created_resource.clone(),
+        keychain.discovery_pk,
+        keychain.encryption_pk,
+    );
+
+    // Verify the transaction
+    assert!(tx.clone().verify(), "Transaction verification failed");
+
+    ProtocolAdapter::Transaction::from(tx)
+}
+
+fn transfer_consumed_created_pair(
+    d: &SetUp,
+) -> (
+    (arm_risc0::resource::Resource, NullifierKey),
+    (arm_risc0::resource::Resource, NullifierKey),
+) {
+    // Obtain the consumed resource data
+    let consumed_auth_sk = AuthorizationSigningKey::new();
+    let consumed_auth_pk = AuthorizationVerifyingKey::from_signing_key(&consumed_auth_sk);
+    let (consumed_nf_key, consumed_nf_cm) = NullifierKey::random_pair();
+
+    let consumed_resource = construct_persistent_resource(
+        &d.spender.to_vec(), // forwarder_addr
+        &d.erc20.to_vec(),   // token_addr
+        d.amount.try_into().unwrap(),
+        vec![4u8; 32], // nonce
+        consumed_nf_cm,
+        vec![5u8; 32], // rand_seed
+        &consumed_auth_pk,
+    );
+    let consumed_nf = consumed_resource.nullifier(&consumed_nf_key).unwrap();
+
+    // Create the ephemeral resource
+    let (created_nf_key, created_nf_cm) = NullifierKey::random_pair();
+    let (_created_discovery_sk, created_discovery_pk) = random_keypair();
+    let created_resource = construct_ephemeral_resource(
+        &d.spender.to_vec(), // forwarder_addr
+        &d.erc20.to_vec(),   // token_addr
+        d.amount.try_into().unwrap(),
+        consumed_nf.as_bytes().to_vec(), // nonce
+        created_nf_cm,
+        vec![6u8; 32], // rand_seed
+        CallType::Unwrap,
+        &d.signer.address().to_vec(), // user_addr
+    );
+
+    (
+        (consumed_resource, consumed_nf_key),
+        (created_resource, created_nf_key),
+    )
+}
+
+fn sha256(a: &[u8], b: &[u8]) -> B256 {
+    let mut hasher = Sha256::new();
+    hasher.update(a);
+    hasher.update(b);
+
+    B256::from_slice(&hasher.finalize())
+}
+
+fn write_to_file(tx: ProtocolAdapter::Transaction, file_name: &str) {
+    use alloy::sol_types::SolValue; // Import the trait for abi_encode
+
+    let encoded_tx = tx.abi_encode();
+
+    std::fs::write(
+        format!("{file_name}.json"),
+        serde_json::to_string_pretty(&tx).unwrap(),
+    )
+    .expect("Failed to write file");
+
+    std::fs::write(format!("{file_name}.bin"), encoded_tx).expect("Failed to write file");
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::conversion::ProtocolAdapter;
+    use std::env;
+
+    #[test]
+    #[ignore]
+    fn print_tx() {
+        env::var("BONSAI_API_KEY").expect("Couldn't read BONSAI_API_KEY");
+        env::var("BONSAI_API_URL").expect("Couldn't read BONSAI_API_URL");
+
+        println!(
+            "{:?}",
+            B256::from_slice(arm_risc0::constants::COMPLIANCE_VK.as_bytes())
+        );
+
+        let n_actions = 1;
+
+        let raw_tx = arm_risc0::transaction::generate_test_transaction(n_actions);
+        println!("{:?}", raw_tx);
+        let evm_tx = ProtocolAdapter::Transaction::from(raw_tx);
+
+        use alloy::sol_types::SolValue;
+        let encoded_tx = evm_tx.abi_encode();
+        let decoded_tx: ProtocolAdapter::Transaction =
+            ProtocolAdapter::Transaction::abi_decode(&encoded_tx).unwrap();
+        assert_eq!(evm_tx, decoded_tx);
+
+        write_to_file(evm_tx, &format!("test_tx{n_actions:02}"));
+    }
+
+    #[test]
+    fn generates_simple_transfer_txns() {
+        env::var("BONSAI_API_KEY").expect("Couldn't read BONSAI_API_KEY");
+        env::var("BONSAI_API_URL").expect("Couldn't read BONSAI_API_URL");
+
+        let d = default_values();
+        let keychain = keychain_a();
+
+        let mint_tx = mint_tx(&d, &keychain);
+        write_to_file(mint_tx, "mint");
+
+        let transfer_tx = transfer_tx(&d, &keychain);
+        write_to_file(transfer_tx, "transfer");
+    }
+}
