+++ bindings/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod call;
 pub mod conversion;
+mod erc20_forwarder;
 pub mod permit2;
