+++ bindings/src/permit2.rs
@@ -1,13 +1,13 @@
 //! Example of how to transfer ERC20 tokens from one account to another using a signed permit.
 
 use alloy::primitives::{
-    address, bytes, keccak256, Address, Bytes, FixedBytes, Signature, B256, U256,
+    bytes, keccak256, Address, Bytes, FixedBytes, Signature, B256, U256,
 };
 use alloy::signers::{local::PrivateKeySigner, Signer};
 use alloy::sol;
 use alloy::sol_types::SolValue;
 
-const CANONICAL_PERMIT2_ADDRESS: Address = address!("0x000000000022D473030F116dDEE9F6B43aC78BA3");
+// const CANONICAL_PERMIT2_ADDRESS: Address = address!("0x000000000022D473030F116dDEE9F6B43aC78BA3");
 
 sol!(
     #[allow(missing_docs)]
@@ -106,31 +106,8 @@ pub fn permit_witness_transfer_from_digest(
     keccak256(data)
 }
 
-pub struct SetUp {
-    pub signer: PrivateKeySigner,
-    pub erc20: Address,
-    pub amount: U256,
-    pub nonce: U256,
-    pub deadline: U256,
-    pub spender: Address,
-}
-
-pub fn default_values() -> SetUp {
-    SetUp {
-        signer: "0x97ecae11e1bd9b504ff977ae3815599331c6b0757ee4af3140fe616adb19ae45"
-            .parse()
-            .expect("should parse private key"),
-        erc20: address!("0x5615dEB798BB3E4dFa0139dFa1b3D433Cc23b72f"),
-        amount: U256::from(1000),
-        nonce: U256::from(0),
-        deadline: U256::from(1789040701),
-        spender: address!("0xA4AD4f68d0b91CFD19687c881e50f3A00242828c"),
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use tokio;
 
     /*
     #[test]
